+++ src/form-validator/field-validator.js
@@ -1,5 +1,4 @@
 import { parseValidatorRules, runRule } from "./validator-rule";
-import { getFieldDisplayData } from "../data-util";
 
 class FieldValidator {
   constructor(field, rules) {
@@ -7,6 +6,10 @@ class FieldValidator {
     this.disabled = false;
     this.initRules(rules);
   }
+  /**
+   * 初始化Rule
+   * @param {Array<Object>} rules rule列表
+   */
   initRules(rules) {
     let list = [];
     rules.forEach(item => {
@@ -14,20 +17,36 @@ class FieldValidator {
     });
     this.rules = list;
   }
+  /**
+   * 禁用字段校验器
+   * @param {Boolean} disable 是否禁用
+   */
   setDisable(disable = true) {
-    this.disabled = true;
+    this.disabled = disable;
   }
-  setRuleDisable(disabled, name) {
-    const isNumber = typeof name === "number";
+  /**
+   * 禁用校验规则
+   * @param {String|Number} rule 校验规则名称或者序号，需要不推荐使用
+   * @param {Boolean} disabled 是否禁用
+   */
+  setRuleDisable(rule, disabled) {
+    const isNumber = typeof rule === "number";
 
-    this.rules.forEach((rule, index) => {
-      if (isNumber && index === name) {
-        rule.disabled = disabled;
-      } else if (rule.name === name) {
-        rule.disabled = disabled;
+    this.rules.forEach((ruleItem, index) => {
+      if (isNumber && index === rule) {
+        ruleItem.disabled = disabled;
+      } else if (rule.name === rule) {
+        ruruleItemle.disabled = disabled;
       }
     });
   }
+  /**
+   * 执行校验规则，返回执行状态
+   * @param {Function} dispatch 校验状态更新器
+   * @param {String} getState 获取当前State的方法
+   * @param {Object} payload 参数，包括：field,vField,formName,data,value等
+   * @returns validateResult 校验结果：{validateStatus,hasFeedback,help}
+   */
   async run(dispatch, getState, payload) {
     let validateResult = null;
     const { rules } = this;